confirm delete popup.

+++ src/_store/users.module.js
@@ -16,13 +16,15 @@ const actions = {
     },
 
     delete({ commit }, id) {
-        commit('deleteRequest', id);
+        if(confirm("Are you sure?")) {
+            commit('deleteRequest', id);
 
-        userService.delete(id)
-            .then(
-                user => commit('deleteSuccess', id),
-                error => commit('deleteFailure', { id, error: error.toString() })
-            );
+            userService.delete(id)
+                .then(
+                    user => commit('deleteSuccess', id),
+                    error => commit('deleteFailure', { id, error: error.toString() })
+                );
+        }
     }
 };
